requirements.txt
flask
flask_sqlalchemy
PyJWT
python-dotenv
pymysql

